LayoutTests/touchadjustment/small-target-test-expected.txt

  
 
 
Tests if small clickable targets adjacent to larger clickable controls are properly targetted with touch adjustment.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Testing small direct hits.
PASS adjustedNode.id is "check"
PASS adjustedNode.id is "task"
PASS adjustedNode.id is "item-label"
PASS adjustedNode.id is "remove-button"
PASS adjustedNode.id is "combo-label"
PASS adjustedNode.id is "drop-down-selector"
Testing indirect hits.
PASS adjustedNode.id is "check"
PASS adjustedNode.id is "check"
PASS adjustedNode.id is "check"
PASS adjustedNode.id is "check"
PASS adjustedNode.id is "check"
PASS adjustedNode.id is "check"
PASS adjustedNode.id is "check"
PASS adjustedNode.id is "check"
PASS adjustedNode.id is "remove-button"
PASS adjustedNode.id is "remove-button"
PASS adjustedNode.id is "remove-button"
PASS adjustedNode.id is "remove-button"
PASS adjustedNode.id is "remove-button"
PASS adjustedNode.id is "remove-button"
PASS adjustedNode.id is "remove-button"
PASS adjustedNode.id is "remove-button"
PASS adjustedNode.id is "drop-down-selector"
PASS adjustedNode.id is "drop-down-selector"
PASS adjustedNode.id is "drop-down-selector"
PASS adjustedNode.id is "drop-down-selector"
PASS adjustedNode.id is "drop-down-selector"
PASS adjustedNode.id is "drop-down-selector"
PASS adjustedNode.id is "drop-down-selector"
PASS adjustedNode.id is "drop-down-selector"
PASS successfullyParsed is true

TEST COMPLETE
